refactor generated types to new types.ts
